Moved entities to the common folder
