+++ src/app/common/entities/Agent.ts
