+++ src/app/common/entities/Flight.ts
